+++ config/management/operational/src/command.rs
@@ -1,7 +1,9 @@
 // Copyright (c) The Libra Core Contributors
 // SPDX-License-Identifier: Apache-2.0
+
+use crate::TransactionContext;
 use libra_crypto::{ed25519::Ed25519PublicKey, x25519};
-use libra_management::{error::Error, TransactionContext};
+use libra_management::error::Error;
 use libra_secure_json_rpc::VMStatusView;
 use libra_types::{validator_config::ValidatorConfig, validator_info::ValidatorInfo};
 use structopt::StructOpt;
