+++ config/management/operational/src/validator_config.rs
@@ -1,15 +1,14 @@
 // Copyright (c) The Libra Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
+use crate::{json_rpc::JsonRpcClientWrapper, TransactionContext};
 use libra_crypto::{ed25519::Ed25519PublicKey, x25519};
 use libra_global_constants::{
     CONSENSUS_KEY, FULLNODE_NETWORK_KEY, OWNER_ACCOUNT, VALIDATOR_NETWORK_KEY,
 };
 use libra_management::{
     error::Error,
-    json_rpc::JsonRpcClientWrapper,
     storage::{to_x25519, StorageWrapper},
-    TransactionContext,
 };
 use libra_network_address::{
     encrypted::{EncNetworkAddress, Key, RawEncNetworkAddress, TEST_SHARED_VAL_NETADDR_KEY},
