+++ config/management/src/lib.rs
@@ -3,10 +3,7 @@
 
 #![forbid(unsafe_code)]
 
-use libra_types::account_address::AccountAddress;
-
 pub mod error;
-pub mod json_rpc;
 pub mod secure_backend;
 pub mod storage;
 pub mod validator_config;
@@ -22,20 +19,4 @@ pub mod constants {
     pub const MAX_GAS_AMOUNT: u64 = 1_000_000;
     pub const GAS_CURRENCY_CODE: &str = LBR_NAME;
     pub const TXN_EXPIRATION_SECS: u64 = 3600;
-}
-
-/// Information for validating a transaction after it's been submitted
-#[derive(Debug, PartialEq)]
-pub struct TransactionContext {
-    pub address: AccountAddress,
-    pub sequence_number: u64,
-}
-
-impl TransactionContext {
-    pub fn new(address: AccountAddress, sequence_number: u64) -> TransactionContext {
-        TransactionContext {
-            address,
-            sequence_number,
-        }
-    }
 }
